chore: composition test
